Move Config classes from executor.py to config.py.
